refactor!: split Registry into registry package + lifecycle helpers
Storage moves into burrow/registry with typed lookup (Get[T],
MustGet[T], GetByName, MustGetByName). Lifecycle orchestration
becomes private helpers inside burrow. burrow.App, burrow.Registry,
burrow.HasDependencies remain as type aliases so downstream code
compiles unchanged.

Driven by repeated XxxFromRegistry helpers in warren and the
God-Object smell in the old Registry type.

Also bumps golang.org/x/net to v0.55.0 (CVE GO-2026-5026, surfaced
by govulncheck on this branch).
