+++ THIRD_PARTY_LICENSES.md
@@ -82,8 +82,8 @@ Generated with [google/go-licenses](https://github.com/google/go-licenses).
 | github.com/x448/float16 | MIT | [LICENSE](https://github.com/x448/float16/blob/v0.8.4/LICENSE) |
 | golang.org/x/crypto | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/crypto/+/v0.51.0:LICENSE) |
 | golang.org/x/mod | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/mod/+/v0.36.0:LICENSE) |
-| golang.org/x/net/idna | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/net/+/v0.53.0:LICENSE) |
-| golang.org/x/sys | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/sys/+/v0.44.0:LICENSE) |
+| golang.org/x/net/idna | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/net/+/v0.55.0:LICENSE) |
+| golang.org/x/sys | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/sys/+/v0.45.0:LICENSE) |
 | golang.org/x/text | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/text/+/v0.37.0:LICENSE) |
 | golang.org/x/time/rate | BSD-3-Clause | [LICENSE](https://cs.opensource.google/go/x/time/+/v0.15.0:LICENSE) |
 | gopkg.in/yaml.v3 | MIT | [LICENSE](https://github.com/go-yaml/yaml/blob/v3.0.1/LICENSE) |
