+++ docs/guide/inter-app-communication.md
@@ -1,28 +1,15 @@
 # Inter-App Communication
 
-Apps can interact with each other through the `Registry` and by declaring dependencies.
-
-## Looking Up an App
-
-Use `Registry.Get()` to retrieve another app by name, then type-assert to access its methods:
+Apps reach sibling apps through the [`registry` package](../reference/server.md#registry). The registry holds every app the server registered with `NewServer`, in order. Apps receive it as `cfg.Registry` inside their `Configure` method:
 
 ```go
 func (a *App) Configure(cfg *burrow.AppConfig, _ *cli.Command) error {
-    authApp, ok := cfg.Registry.Get("auth")
-    if !ok {
-        return fmt.Errorf("auth app not registered")
-    }
-
-    // Type-assert to access the auth app's repository.
-    if provider, ok := authApp.(interface{ Repo() *auth.Repository[auth.EmptyProfile] }); ok {
-        a.authRepo = provider.Repo()
-    }
-
-    return nil
+    // cfg.Registry is the *registry.Registry held by the server.
+    // ...
 }
 ```
 
-This pattern is used by the `admin` app to discover its `AdminAuth` middleware provider from the registry.
+`burrow.Registry` is a type alias for `*registry.Registry` — both names refer to the same type.
 
 ## Declaring Dependencies
 
@@ -34,10 +21,74 @@ func (a *App) Dependencies() []string {
 }
 ```
 
-If a dependency is missing when `NewServer` processes your app, it panics at startup with a clear error message — a programming mistake caught immediately.
+`registry.Add` panics at startup if a declared dependency is missing — a programming mistake caught immediately. `NewServer` also auto-sorts apps so you can pass them in any order.
 
-!!! info "Auto-sorting"
-    `NewServer` automatically sorts apps by their `HasDependencies` declarations — you can list them in any order. If a dependency is missing entirely, the server panics at startup with a clear error message.
+A declared dependency is the foundation for the **Hard-Dependency** lookup pattern below: once `Dependencies()` lists a provider, the registry guarantees it is present, and a typed lookup can panic on programmer error instead of returning one.
+
+## Choosing a Lookup Pattern
+
+| Failure mode | Function | When to use |
+|---|---|---|
+| Panic on missing or wrong type | `registry.MustGet[T](reg)` | The provider is declared in `Dependencies()`. |
+| Returns `(T, bool)` | `registry.Get[T](reg)` | The provider may be absent and your code degrades gracefully. |
+| Iterate, type-assert against an interface | `registry.Apps(reg)` + loop | Any of several apps might implement the contract you need. |
+| Lookup by name, then assert | `registry.GetByName(reg, name)` | You know the exact app name but want a non-typed handle (rare). |
+
+### Hard-Dependency — `registry.MustGet[T]`
+
+Use this when the provider is declared in `Dependencies()`. `MustGet` panics when no app of type `T` is registered, when more than one is, or when the registered app has a different concrete type — all three are programming bugs caught at boot. A downstream project's `pages` app reaching its declared `audit` and `site` providers:
+
+```go
+import "github.com/oliverandrich/burrow/registry"
+
+func (a *App) Configure(cfg *burrow.AppConfig, _ *cli.Command) error {
+    a.audit = registry.MustGet[*audit.App](cfg.Registry).Repo()
+    a.site  = registry.MustGet[*site.App](cfg.Registry)
+    return nil
+}
+```
+
+The type parameter is the lookup key; a wrong type fails at compile time.
+
+### Optional-Service — `registry.Get[T]`
+
+Use this when the provider may not be registered and your code can return `nil` or skip the work. `Get` returns the zero value and `false` when no matching app exists. `sse.BrokerFromRegistry` is the canonical example — it lets a background job publish events when SSE is registered and no-op otherwise:
+
+```go
+func BrokerFromRegistry(reg *burrow.Registry) *EventBroker {
+    if app, ok := registry.Get[*App](reg); ok {
+        return app.Broker()
+    }
+    return nil
+}
+```
+
+### Provider-Discovery — `registry.Apps` + interface assertion
+
+Use this when any number of apps might implement the contract you depend on, and you want to find one (or check there's exactly one) without coupling to concrete types. `contrib/admin` uses this to find an `AdminAuth` provider:
+
+```go
+import "github.com/oliverandrich/burrow/registry"
+
+// burrow.AdminAuth (defined in burrow.go) exposes RequireAuth /
+// RequireStaff / RequireAdmin middleware constructors. contrib/auth
+// implements it; a custom auth contrib could too.
+func (a *App) Configure(cfg *burrow.AppConfig, _ *cli.Command) error {
+    for _, app := range registry.Apps(cfg.Registry) {
+        if aa, ok := app.(burrow.AdminAuth); ok {
+            a.authMiddleware = aa
+            return nil
+        }
+    }
+    return fmt.Errorf("admin: no AdminAuth provider found")
+}
+```
+
+This pattern survives someone swapping `contrib/auth` for a custom provider — the consumer only depends on the interface, not the implementer.
+
+### Name lookup — `registry.GetByName`
+
+`registry.GetByName(reg, "name")` returns the named app as `burrow.App`. Use it when you genuinely need the name as the lookup key (rare) — for example, a maintenance command that toggles a feature on a specific app instance. `MustGetByName` is the panicking equivalent.
 
 ## Using Auth Context
 
@@ -55,6 +106,8 @@ func (a *App) List(w http.ResponseWriter, r *http.Request) error {
 }
 ```
 
+Route-level guards live on the auth app itself: `auth.RequireAuth()`, `auth.RequireStaff()`, and `auth.RequireAdmin()`. See [auth guide](../contrib/auth.md).
+
 ## Using Session Data
 
 Read and write session values from any app:
@@ -62,46 +115,15 @@ Read and write session values from any app:
 ```go
 import "github.com/oliverandrich/burrow/contrib/session"
 
-// Read a value.
 userID := session.GetInt64(r, "user_id")
 locale := session.GetString(r, "locale")
 
-// Write a value (immediately writes the cookie).
 session.Set(w, r, "theme", "dark")
-
-// Remove a value.
 session.Delete(w, r, "theme")
-
-// Clear the entire session.
 session.Clear(w, r)
 ```
 
-## Using the SSE Broker from Non-HTTP Code
+## Related
 
-The SSE broker is normally accessed via `sse.Broker(ctx)` in request handlers. For background jobs and other non-HTTP code, use `sse.BrokerFromRegistry`:
-
-```go
-import "github.com/oliverandrich/burrow/contrib/sse"
-
-func (h *MyJobHandler) Run(ctx context.Context) error {
-    broker := sse.BrokerFromRegistry(h.registry)
-    if broker == nil {
-        return nil // SSE not available
-    }
-
-    broker.Publish("updates", sse.Event{
-        Data: "<span>Job finished</span>",
-    })
-    return nil
-}
-```
-
-This avoids manual type assertions on the registry entry and returns `nil` safely when the SSE app is missing or not yet configured.
-
-## Patterns
-
-**Repository sharing** — expose a `Repo()` method on your app and let other apps access it via type assertion through the registry.
-
-**Middleware guards** — use `auth.RequireAuth()` and `auth.RequireAdmin()` in your route groups to protect endpoints.
-
-**Context values** — read shared data (nav items, CSRF token, locale) from the request context using the provided helpers. See [Context Helpers](../reference/context-helpers.md).
+- [Context Helpers](../reference/context-helpers.md) — reading shared data (nav items, CSRF token, locale) from the request context
+- [Server reference: Registry](../reference/server.md#registry) — the registry API surface and how it fits the boot sequence
